+++ delete_test.go
@@ -6,14 +6,15 @@ import (
 )
 
 func TestDeleteWithWhereMySQL(t *testing.T) {
-	query, args := MySQLQuoted.Delete().
+	query, args := Delete().
+		Dialect(MySQL).
 		From("customers").
 		Where("id", "= ?", 9).
 		Build()
 
-	expectedQuery := "DELETE FROM `customers`\n WHERE (`id` = ?)"
+	expectedQuery := "DELETE FROM customers\n WHERE (id = ?)"
 	if query != expectedQuery {
-		t.Errorf("bad query: %s", query)
+		t.Errorf("bad query: %q", query)
 	}
 
 	expectedArgs := []interface{}{9}
@@ -23,15 +24,15 @@ func TestDeleteWithWhereMySQL(t *testing.T) {
 }
 
 func TestDeleteWithWherePostgres(t *testing.T) {
-	query, args := Postgres.Delete().
+	query, args := Delete().
+		Dialect(Postgres).
 		From("customers").
 		Where("id", "= ?", 9).
 		Build()
 
-	expectedQuery := `DELETE FROM "customers"
- WHERE ("id" = $1)`
+	expectedQuery := "DELETE FROM customers\n WHERE (id = $1)"
 	if query != expectedQuery {
-		t.Errorf("bad query: %s", query)
+		t.Errorf("bad query: %q", query)
 	}
 
 	expectedArgs := []interface{}{9}