+++ update_test.go
@@ -6,15 +6,16 @@ import (
 )
 
 func TestUpdateMySQL(t *testing.T) {
-	query, args := MySQLQuoted.Update().
+	query, args := Update().
+		Dialect(MySQL).
 		Table("customers").
 		Set("name", "John").
 		Set("phone", "555").
 		Build()
 
-	expectedQuery := "UPDATE `customers` SET `name` = ?, `phone` = ?"
+	expectedQuery := "UPDATE customers SET name = ?, phone = ?"
 	if query != expectedQuery {
-		t.Errorf("bad query: %s", query)
+		t.Errorf("bad query: %q", query)
 	}
 
 	expectedArgs := []interface{}{"John", "555"}
@@ -24,15 +25,16 @@ func TestUpdateMySQL(t *testing.T) {
 }
 
 func TestUpdatePostgres(t *testing.T) {
-	query, args := Postgres.Update().
+	query, args := Update().
+		Dialect(Postgres).
 		Table("customers").
 		Set("name", "John").
 		Set("phone", "555").
 		Build()
 
-	expectedQuery := `UPDATE "customers" SET "name" = $1, "phone" = $2`
+	expectedQuery := `UPDATE customers SET name = $1, phone = $2`
 	if query != expectedQuery {
-		t.Errorf("bad query: %s", query)
+		t.Errorf("bad query: %q", query)
 	}
 
 	expectedArgs := []interface{}{"John", "555"}
@@ -42,7 +44,8 @@ func TestUpdatePostgres(t *testing.T) {
 }
 
 func TestUpdateWithWhereMySQL(t *testing.T) {
-	query, args := MySQLQuoted.Update().
+	query, args := Update().
+		Dialect(MySQL).
 		Table("customers").
 		Set("name", "John").
 		Set("phone", "555").
@@ -50,9 +53,9 @@ func TestUpdateWithWhereMySQL(t *testing.T) {
 		Where("name", "NOT NULL").
 		Build()
 
-	expectedQuery := "UPDATE `customers` SET `name` = ?, `phone` = ?\n WHERE (`id` = ?) AND (`name` NOT NULL)"
+	expectedQuery := "UPDATE customers SET name = ?, phone = ?\n WHERE (id = ?) AND (name NOT NULL)"
 	if query != expectedQuery {
-		t.Errorf("bad query: %s", query)
+		t.Errorf("bad query: %q", query)
 	}
 
 	expectedArgs := []interface{}{"John", "555", 9}
@@ -62,17 +65,17 @@ func TestUpdateWithWhereMySQL(t *testing.T) {
 }
 
 func TestUpdateWithWherePostgres(t *testing.T) {
-	query, args := Postgres.Update().
+	query, args := Update().
+		Dialect(Postgres).
 		Table("customers").
 		Set("name", "John").
 		Set("phone", "555").
 		Where("id", "= ?", 9).
 		Build()
 
-	expectedQuery := `UPDATE "customers" SET "name" = $1, "phone" = $2
- WHERE ("id" = $3)`
+	expectedQuery := "UPDATE customers SET name = $1, phone = $2\n WHERE (id = $3)"
 	if query != expectedQuery {
-		t.Errorf("bad query: %s", query)
+		t.Errorf("bad query: %q", query)
 	}
 
 	expectedArgs := []interface{}{"John", "555", 9}
@@ -82,12 +85,12 @@ func TestUpdateWithWherePostgres(t *testing.T) {
 }
 
 func TestUpdateReuse(t *testing.T) {
-	baseStatement := MySQL.Update().Table("customers").Set("name", "John")
+	baseStatement := Update().Dialect(MySQL).Table("customers").Set("name", "John")
 
 	query, args := baseStatement.Set("phone", "555").Build()
 	expectedQuery := "UPDATE customers SET name = ?, phone = ?"
 	if query != expectedQuery {
-		t.Errorf("bad query: %s", query)
+		t.Errorf("bad query: %q", query)
 	}
 	expectedArgs := []interface{}{"John", "555"}
 	if !reflect.DeepEqual(args, expectedArgs) {
@@ -97,7 +100,7 @@ func TestUpdateReuse(t *testing.T) {
 	query, args = baseStatement.Set("city", "Berlin").Build()
 	expectedQuery = "UPDATE customers SET name = ?, city = ?"
 	if query != expectedQuery {
-		t.Errorf("bad query: %s", query)
+		t.Errorf("bad query: %q", query)
 	}
 	expectedArgs = []interface{}{"John", "Berlin"}
 	if !reflect.DeepEqual(args, expectedArgs) {