+++ where.go
@@ -10,20 +10,12 @@ type where struct {
 	args     []interface{}
 }
 
-func quoteStrings(s []string, dialect Dialect) []string {
-	var quoted []string
-	for _, o := range s {
-		quoted = append(quoted, dialect.Quote(o))
-	}
-	return quoted
-}
-
 func buildWhereClause(query string, args []interface{}, idx int, wheres []where, dialect Dialect) (string, []interface{}, int) {
 	if len(wheres) > 0 {
 		var sqls []string
 
 		for _, where := range wheres {
-			sql := "(" + dialect.Quote(where.col) + " " + where.sql + ")"
+			sql := "(" + where.col + " " + where.sql + ")"
 			for _, arg := range where.args {
 				value := reflect.ValueOf(arg)
 				switch value.Kind() {