+++ CHANGES.md
@@ -1,5 +1,13 @@
 # Changes
 
+## 3.0.0
+
+* Remove quoting of table and column names introduced in 2.0.0
+  (as a consequence, the `MapSQL`, `ReturnSQL`, and `Quote` methods are gone)
+* Introduce dialects (replacing DBMS)
+
+Version 3 is not compatible with version 1 and 2.
+
 ## 2.0.0
 
 * Add support for `RETURNING` clause